The module configuration import/export works \o/ Signed-off-by: Tim Strijdhorst <t.strijdhorst@student.utwente.nl>

+++ src/GenerationFramework/src/org/hazelwire/gui/TabItemTest.java
@@ -18,7 +18,7 @@ public class TabItemTest extends TabItem implements Drawable
 	}
 
 	@Override
-	public int internal_new_GC(GCData arg0)
+	public long internal_new_GC(GCData arg0)
 	{
 		return 0;
 	}
